backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
python-decouple==3.8
django-filter==23.5
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0

# LCA and Environmental Impact Assessment
brightway2==2.5
bw2data==4.0.2
bw2calc==1.8.0
bw2io==0.8.5
bw2parameters==0.6.6
bw2analyzer==0.9.4
bw2speedups==0.3.5
lxml==4.9.3
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1













